requirements.txt
fastapi
uvicorn
tabula-py
pandas
sqlalchemy
pyodbc
PyPDF2
